Trident and Crossbow Enchantents Support
Added #luckyTridentEnchantments and, #luckyCrossbowEnchantments which both give an item Trident and Crossbows respectively as long as they use Enchantments=#luckyTridentEnchantments, as an example. This also adds Trident and Crossbow Enchantments to #randEnchantment which means there's a possibility of getting an enchanted book with Trident or Crossbow enchantments in the base lucky block mod.